[ALPS04776785] ion: correct iova addr
[Detail]
correct iova addr from unsigned int to unsigned long

MTK-Commit-Id: fe986426784268025f153f37143dcebbc26b6b32

CR-Id: ALPS04776785
Feature: [Module]ION/M4U

+++ drivers/staging/android/ion/mtk/ion_drv.c
@@ -556,7 +556,7 @@ static long ion_sys_ioctl(struct ion_client *client, unsigned int cmd,
 				       __func__);
 				ret = -EFAULT;
 			}
-			param.get_phys_param.phy_addr = (unsigned int)phy_addr;
+			param.get_phys_param.phy_addr = phy_addr;
 			ion_drv_put_kernel_handle(kernel_handle);
 		}
 		break;
